[REFACTOR] stubs now gets imports provided via set, rather than extracting them itself.
