+++ lib/stub-generator.js
@@ -5,6 +5,7 @@ var Stubs = require('./stubs');
 var fs = require('fs');
 var md5Hex = require('md5-hex');
 var debug = require('debug')('ember-browserify:stub-generator:');
+var importsFor = require('./imports-for');
 
 module.exports = StubGenerator;
 function StubGenerator(inputTree, options) {
@@ -54,7 +55,10 @@ StubGenerator.prototype.build = function() {
     switch (operation) {
       case 'unlink': this.stubs.delete(fullPath); break;
       case 'create':
-      case 'change': this.stubs.set(fullPath, fs.readFileSync(fullPath)); break;
+      case 'change':
+        var content = fs.readFileSync(fullPath, 'UTF8');
+        this.stubs.set(fullPath, importsFor(content));
+        break;
     }
   }, this);
